Moved all components to the components dir
